Fix migration error

+++ database/migrations/2016_03_21_095359_create_project_hyperlinks_table.php
@@ -13,9 +13,9 @@ class CreateProjectHyperlinksTable extends Migration
     public function up()
     {
         Schema::create('project_hyperlinks', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->integer('project_id')->unsigned();
-            $table->integer('hyperlink_id')->unsigned();
+            $table->bigInteger('hyperlink_id')->unsigned();
 
             $table->foreign('project_id')->references('id')
                 ->on('projects')->onDelete('cascade')
